Add Cached Shader Compile & Fix Rendering
Changelog:
* Added cacheing to cmake script that compiles SPIRV & .NYX files, so
builds after the first one aren't terrible.
* Added RenderTest.cpp for testing rendering multiple frames. RenderDoc
does not capture the normal tests.
* Fixed bug with render pass not generating framebuffers correctly from
swapchain images.
* Updated .NYX file spec. Prob should be a big deal, but no one uses it
besides me right now so! It's ok.
* Added image handling of depth images. Now, if you pass a depth
layout/format to an image, it knows it's supposed to be used for depth
and does all the vulkan things it needs to do.
* Fixed CommandBuffer submittion & handling. It now correctly uses the
render pass for all indexing.
* CommandBuffer now functions as a Vector of command buffers + an
iterator for the 'head'.
* CommandBuffers now have fences for each command buffer, so now submits
can happen on the same command buffer and it automatically waits on the
fence in case the buffer was previously in use by the GPU.
* All Vulkan Memory Buffers now have TransferSrc and TransferDst flags
by default. Why would you never want these?
* Added VKG library errors to catch.
* Lots of other things but I can't remember. Big changelists are bad.